requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==22.0.0
packaging==24.1
PyJWT==2.8.0
pyotp==2.9.0
pypng==0.20220715.0
python-decouple==3.8
pytz==2024.1
qrcode==7.4.2
sqlparse==0.5.0
typing_extensions==4.12.2
